edms-production-20260106-170206/backend/requirements/base.txt
# Core Django packages
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.7
# django-pgvector==0.2.1  # Removed - not compatible with Python 3.11

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7

# Workflow Engine
django-viewflow>=2.2.13

# File Processing
python-docx==1.1.0
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.0.1

# Background Tasks & Cache
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# WebSocket Support removed - HTTP polling used instead

# API Documentation
drf-spectacular==0.26.5

# Development Tools
django-debug-toolbar==4.2.0
ipython==8.16.1

# Monitoring & Logging
sentry-sdk[django]==1.35.0

# Environment Management
python-decouple==3.8  # Phase 3 High Priority Dependencies

# Additional Document Processing
reportlab==4.0.4
weasyprint==60.2
qrcode==7.4.2
docxtpl==0.20.2

# Electronic Signatures
cryptography==41.0.7

# Scheduler Enhancement
django-celery-beat==2.5.0
pytz==2024.1
